bpf: Fix maglev hash with hostServices.hostNamespaceOnly
[ upstream commit fc99c0cdfa37b7da7b27405deef68dbef2185c5d ]

This fixes the bug that Cilium drops packets destined to ClusterIP.
Cilium tries to resolve ClusterIP to Pod IP using Maglev hash when
hostServices.hostNamespaceOnly is enabled. However, it doesn't provision
the Maglev LUT for ClusterIP. So it drops the packet.
This commit fixes this problem by forcing bpf_lxc to use the random
backend selection for ClusterIP. Also, Cilium will populate the Maglev
LUT for ClusterIP if bpf.lbExternalClusterIP is set to true so that
the external ingress traffic connecting ClusterIP will be properly
handled.

Fixes: #17474

[ Backport notes: Minor conflicts in test/k8sT/Services.go; the second
    hunk in that file was skipped (the redeploy is not present in this
    branch). ]

Signed-off-by: Yusuke Suzuki <yusuke-suzuki@cybozu.co.jp>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>